refactor: 불필요한 코드 삭제 및 변수명 변경
