+++ src/App.jsx
@@ -1,0 +1,7 @@
+import CardAppPage from './page/CardAddPage';
+
+const App = () => {
+  return <CardAppPage />;
+};
+
+export default App;
