+++ src/components/CardCompany.jsx
@@ -1,10 +1,8 @@
-import React from 'react';
-
-const CardCompany = ({ name, onClickCompany, theme }) => {
+const CardCompany = ({ company, onClickCompany, theme }) => {
   return (
-    <div className="modal-item-container" onClick={() => onClickCompany(name, theme)}>
+    <div className="modal-item-container" onClick={() => onClickCompany(company, theme)}>
       <div className="modal-item-dot" style={{ backgroundColor: theme }}></div>
-      <span className="modal-item-name">{name}</span>
+      <span className="modal-item-name">{company}</span>
     </div>
   );
 };
