+++ src/components/common/Modal.jsx
@@ -1,5 +1,3 @@
-import React from 'react';
-
 const Modal = ({ handleModal, children }) => {
   return (
     <>
